Added basic implementation of reposcan based on connexion
This uses similar approach to the one used in vulnerability engine, with individual API calls being
implemented in classes, with basic behavior implemented in base classes from base.py. The server uses
WSGI container running on tornado, which contains the application, since we require the IOLoop for
current implementation of websocket client.