Reduce number of checked packages in /updates logic
This commit changes the way we encode updates, so that there are separate lists for
each architecture instead of one single list. This should reduce amount of packages
we need to traverse in order to find valid updates, and therefore reduce /updates
call latency.

There might be some duplication, since we store duplicate information (noarch packages)
but thius should be on the order of several MBs, maybe tens...

The order of updates for noarch packages might also be slightly altered